+++ homeassistant/components/ecobee/__init__.py
@@ -27,6 +27,7 @@ from homeassistant.helpers.typing import ConfigType
 from homeassistant.util import Throttle
 
 from .const import _LOGGER, CONF_REFRESH_TOKEN, DOMAIN, PLATFORMS
+from .services import async_setup_services
 
 MIN_TIME_BETWEEN_UPDATES = timedelta(seconds=180)
 
@@ -36,11 +37,8 @@ type EcobeeConfigEntry = ConfigEntry[EcobeeData]
 
 
 async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
-    """Set up ecobee."""
-    from .climate import async_register_services  # noqa: PLC0415
-
-    async_register_services(hass)
-
+    """Set up the ecobee integration."""
+    async_setup_services(hass)
     return True
