+++ homeassistant/components/ecobee/climate.py
@@ -21,7 +21,6 @@ from homeassistant.components.climate import (
     HVACMode,
 )
 from homeassistant.const import (
-    ATTR_ENTITY_ID,
     ATTR_TEMPERATURE,
     PRECISION_HALVES,
     PRECISION_TENTHS,
@@ -29,7 +28,7 @@ from homeassistant.const import (
     STATE_ON,
     UnitOfTemperature,
 )
-from homeassistant.core import HomeAssistant, ServiceCall, callback
+from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ServiceValidationError
 from homeassistant.helpers import (
     config_validation as cv,
@@ -50,18 +49,20 @@ from .const import (
     ECOBEE_MODEL_TO_NAME,
     MANUFACTURER,
 )
-from .util import ecobee_date, ecobee_time, is_indefinite_hold
+from .services import (
+    ATTR_COOL_TEMP,
+    ATTR_END_DATE,
+    ATTR_END_TIME,
+    ATTR_FAN_MIN_ON_TIME,
+    ATTR_FAN_MODE,
+    ATTR_HEAT_TEMP,
+    ATTR_START_DATE,
+    ATTR_START_TIME,
+    ATTR_VACATION_NAME,
+    _async_get_thermostats,
+)
+from .util import is_indefinite_hold
 
-ATTR_COOL_TEMP = "cool_temp"
-ATTR_END_DATE = "end_date"
-ATTR_END_TIME = "end_time"
-ATTR_FAN_MIN_ON_TIME = "fan_min_on_time"
-ATTR_FAN_MODE = "fan_mode"
-ATTR_HEAT_TEMP = "heat_temp"
-ATTR_RESUME_ALL = "resume_all"
-ATTR_START_DATE = "start_date"
-ATTR_START_TIME = "start_time"
-ATTR_VACATION_NAME = "vacation_name"
 ATTR_DST_ENABLED = "dst_enabled"
 ATTR_MIC_ENABLED = "mic_enabled"
 ATTR_AUTO_AWAY = "auto_away"
@@ -69,7 +70,6 @@ ATTR_FOLLOW_ME = "follow_me"
 ATTR_SENSOR_LIST = "device_ids"
 ATTR_PRESET_MODE = "preset_mode"
 
-DEFAULT_RESUME_ALL = False
 PRESET_AWAY_INDEFINITELY = "away_indefinitely"
 PRESET_TEMPERATURE = "temp"
 PRESET_VACATION = "vacation"
@@ -129,68 +129,10 @@ PRESET_TO_ECOBEE_HOLD = {
     PRESET_HOLD_INDEFINITE: "indefinite",
 }
 
-SERVICE_CREATE_VACATION = "create_vacation"
-SERVICE_DELETE_VACATION = "delete_vacation"
-SERVICE_RESUME_PROGRAM = "resume_program"
-SERVICE_SET_FAN_MIN_ON_TIME = "set_fan_min_on_time"
 SERVICE_SET_DST_MODE = "set_dst_mode"
 SERVICE_SET_MIC_MODE = "set_mic_mode"
 SERVICE_SET_OCCUPANCY_MODES = "set_occupancy_modes"
 SERVICE_SET_SENSORS_USED_IN_CLIMATE = "set_sensors_used_in_climate"
-
-DTGROUP_START_INCLUSIVE_MSG = (
-    f"{ATTR_START_DATE} and {ATTR_START_TIME} must be specified together"
-)
-
-DTGROUP_END_INCLUSIVE_MSG = (
-    f"{ATTR_END_DATE} and {ATTR_END_TIME} must be specified together"
-)
-
-CREATE_VACATION_SCHEMA = vol.Schema(
-    {
-        vol.Required(ATTR_ENTITY_ID): cv.entity_id,
-        vol.Required(ATTR_VACATION_NAME): vol.All(cv.string, vol.Length(max=12)),
-        vol.Required(ATTR_COOL_TEMP): vol.Coerce(float),
-        vol.Required(ATTR_HEAT_TEMP): vol.Coerce(float),
-        vol.Inclusive(
-            ATTR_START_DATE, "dtgroup_start", msg=DTGROUP_START_INCLUSIVE_MSG
-        ): ecobee_date,
-        vol.Inclusive(
-            ATTR_START_TIME, "dtgroup_start", msg=DTGROUP_START_INCLUSIVE_MSG
-        ): ecobee_time,
-        vol.Inclusive(
-            ATTR_END_DATE, "dtgroup_end", msg=DTGROUP_END_INCLUSIVE_MSG
-        ): ecobee_date,
-        vol.Inclusive(
-            ATTR_END_TIME, "dtgroup_end", msg=DTGROUP_END_INCLUSIVE_MSG
-        ): ecobee_time,
-        vol.Optional(ATTR_FAN_MODE, default="auto"): vol.Any("auto", "on"),
-        vol.Optional(ATTR_FAN_MIN_ON_TIME, default=0): vol.All(
-            int, vol.Range(min=0, max=60)
-        ),
-    }
-)
-
-DELETE_VACATION_SCHEMA = vol.Schema(
-    {
-        vol.Required(ATTR_ENTITY_ID): cv.entity_id,
-        vol.Required(ATTR_VACATION_NAME): vol.All(cv.string, vol.Length(max=12)),
-    }
-)
-
-RESUME_PROGRAM_SCHEMA = vol.Schema(
-    {
-        vol.Optional(ATTR_ENTITY_ID): cv.entity_ids,
-        vol.Optional(ATTR_RESUME_ALL, default=DEFAULT_RESUME_ALL): cv.boolean,
-    }
-)
-
-SET_FAN_MIN_ON_TIME_SCHEMA = vol.Schema(
-    {
-        vol.Optional(ATTR_ENTITY_ID): cv.entity_ids,
-        vol.Required(ATTR_FAN_MIN_ON_TIME): vol.Coerce(int),
-    }
-)
 
 
 SUPPORT_FLAGS = (
@@ -199,104 +141,6 @@ SUPPORT_FLAGS = (
     | ClimateEntityFeature.TARGET_TEMPERATURE_RANGE
     | ClimateEntityFeature.FAN_MODE
 )
-
-
-@callback
-def async_register_services(hass: HomeAssistant) -> None:
-    """Register ecobee services."""
-    # pylint: disable-next=home-assistant-use-runtime-data
-    hass.data.setdefault(DOMAIN, {})[DATA_THERMOSTATS] = []
-
-    def create_vacation_service(service: ServiceCall) -> None:
-        """Create a vacation on the target thermostat."""
-        entity_id = service.data[ATTR_ENTITY_ID]
-
-        for thermostat in _async_get_thermostats(hass):
-            if thermostat.entity_id == entity_id:
-                thermostat.create_vacation(service.data)
-                thermostat.schedule_update_ha_state(True)
-                break
-
-    def delete_vacation_service(service: ServiceCall) -> None:
-        """Delete a vacation on the target thermostat."""
-        entity_id = service.data[ATTR_ENTITY_ID]
-        vacation_name = service.data[ATTR_VACATION_NAME]
-
-        for thermostat in _async_get_thermostats(hass):
-            if thermostat.entity_id == entity_id:
-                thermostat.delete_vacation(vacation_name)
-                thermostat.schedule_update_ha_state(True)
-                break
-
-    def fan_min_on_time_set_service(service: ServiceCall) -> None:
-        """Set the minimum fan on time on the target thermostats."""
-        entity_id = service.data.get(ATTR_ENTITY_ID)
-        fan_min_on_time = service.data[ATTR_FAN_MIN_ON_TIME]
-
-        thermostats = _async_get_thermostats(hass)
-        if entity_id:
-            target_thermostats = [
-                thermostat
-                for thermostat in thermostats
-                if thermostat.entity_id in entity_id
-            ]
-        else:
-            target_thermostats = thermostats
-
-        for thermostat in target_thermostats:
-            thermostat.set_fan_min_on_time(str(fan_min_on_time))
-            thermostat.schedule_update_ha_state(True)
-
-    def resume_program_set_service(service: ServiceCall) -> None:
-        """Resume the program on the target thermostats."""
-        entity_id = service.data.get(ATTR_ENTITY_ID)
-        resume_all = service.data.get(ATTR_RESUME_ALL)
-
-        thermostats = _async_get_thermostats(hass)
-        if entity_id:
-            target_thermostats = [
-                thermostat
-                for thermostat in thermostats
-                if thermostat.entity_id in entity_id
-            ]
-        else:
-            target_thermostats = thermostats
-
-        for thermostat in target_thermostats:
-            thermostat.resume_program(resume_all)
-            thermostat.schedule_update_ha_state(True)
-
-    hass.services.async_register(
-        DOMAIN,
-        SERVICE_CREATE_VACATION,
-        create_vacation_service,
-        schema=CREATE_VACATION_SCHEMA,
-    )
-    hass.services.async_register(
-        DOMAIN,
-        SERVICE_DELETE_VACATION,
-        delete_vacation_service,
-        schema=DELETE_VACATION_SCHEMA,
-    )
-    hass.services.async_register(
-        DOMAIN,
-        SERVICE_SET_FAN_MIN_ON_TIME,
-        fan_min_on_time_set_service,
-        schema=SET_FAN_MIN_ON_TIME_SCHEMA,
-    )
-    hass.services.async_register(
-        DOMAIN,
-        SERVICE_RESUME_PROGRAM,
-        resume_program_set_service,
-        schema=RESUME_PROGRAM_SCHEMA,
-    )
-
-
-@callback
-def _async_get_thermostats(hass: HomeAssistant) -> list[Thermostat]:
-    """Return loaded ecobee thermostat entities."""
-    # pylint: disable-next=home-assistant-use-runtime-data
-    return hass.data[DOMAIN][DATA_THERMOSTATS]
 
 
 async def async_setup_entry(
